add life system, that is trigger to translate result state

+++ src/component/player.rs
@@ -1,4 +1,4 @@
 #[derive(Default)]
 pub struct Player {
-    pub life: usize,
+    pub life: i32,
 }
